+++ src/actions/commit_create.ts
@@ -1,7 +1,7 @@
 import { TContext } from '../lib/context';
+import { SCOPE } from '../lib/engine/scope_spec';
 import { addAll } from '../lib/git/add_all';
 import { ensureSomeStagedChangesPrecondition } from '../lib/preconditions';
-import { SCOPE } from '../lib/state/scope_spec';
 import { restackBranches } from './restack';
 
 export function commitCreateAction(
