+++ src/actions/edit/edit_downstack.ts
@@ -1,6 +1,6 @@
 import { TContext } from '../../lib/context';
+import { SCOPE } from '../../lib/engine/scope_spec';
 import { ExitFailedError } from '../../lib/errors';
-import { SCOPE } from '../../lib/state/scope_spec';
 import { getDefaultEditorOrPrompt } from '../../lib/utils/default_editor';
 import { gpExecSync } from '../../lib/utils/exec_sync';
 import { performInTmpDir } from '../../lib/utils/perform_in_tmp_dir';
