+++ src/actions/show_branch.ts
@@ -1,8 +1,8 @@
 import chalk from 'chalk';
 import { TContext } from '../lib/context';
+import { TBranchPRInfo } from '../lib/engine/metadata_ref';
 import { getCommitterDate } from '../lib/git/committer_date';
 import { showCommits } from '../lib/git/show_commits';
-import { TBranchPRInfo } from '../lib/state/metadata_ref';
 
 export async function showBranchAction(
   branchName: string,
