+++ src/actions/submit/submit_action.ts
@@ -3,9 +3,9 @@ import * as t from '@withgraphite/retype';
 import chalk from 'chalk';
 import prompts from 'prompts';
 import { TContext } from '../../lib/context';
+import { TScopeSpec } from '../../lib/engine/scope_spec';
 import { KilledError } from '../../lib/errors';
 import { cliAuthPrecondition } from '../../lib/preconditions';
-import { TScopeSpec } from '../../lib/state/scope_spec';
 import { getSurvey, showSurvey } from '../../lib/telemetry/survey/survey';
 import { Unpacked } from '../../lib/utils/ts_helpers';
 import { getPRInfoForBranches } from './prepare_branches';
