+++ src/commands/upstack-commands/restack.ts
@@ -1,6 +1,6 @@
 import yargs from 'yargs';
 import { restackBranches } from '../../actions/restack';
-import { SCOPE } from '../../lib/state/scope_spec';
+import { SCOPE } from '../../lib/engine/scope_spec';
 import { profile } from '../../lib/telemetry/profile';
 
 const args = {} as const;
