+++ src/lib/engine/scope_spec.ts
